requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
gspread>=6.0.0
google-auth>=2.20.0
